requirements.txt
fastapi
uvicorn
pyswisseph
requests
timezonefinder
pytz
streamlit
google-generativeai
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart